refactor(agent-toolkit): remove deprecated form tools and schemas
- Deleted obsolete tools: CreateFormTool, FormQuestionsEditorTool, GetFormTool, UpdateFormTool, and associated schemas to streamline the agent toolkit.
- This cleanup enhances maintainability and focuses on current functionalities, ensuring a more efficient codebase.
