+++ packages/agent-toolkit/src/core/tools/platform-api-tools/workforms-tools/get-form-tool/schema.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+import { GraphQLDescriptions } from '../workforms.consts';
+
+export const getFormToolSchema = {
+  formToken: z.string().describe(GraphQLDescriptions.commonArgs.formToken),
+};
